+++ example/peripheral/adc/sdkconfig.h
@@ -159,6 +159,11 @@
 
 #define CONFIG_FREERTOS_USE_ADC
 /* end of Freertos Adc Drivers */
+
+/* Freertos Can Drivers */
+
+/* CONFIG_FREERTOS_USE_CAN is not set */
+/* end of Freertos Can Drivers */
 /* end of Component Configuration */
 
 /* FreeRTOS Setting */